graphragREC/requirements.txt
future>=1.0.0
openai>=0.27.0
tiktoken
networkx
graspologic
nano-vectordb
hnswlib
xxhash
tenacity
dspy-ai
neo4j
aioboto3

# 추가 패키지
python-dotenv>=1.0.0
nest_asyncio>=1.5.6
sentence-transformers>=2.2.2
torch>=1.12.0
numpy>=1.21.0
pandas>=1.3.0
